requirements.txt
firebase_admin
fastapi
requests
uvicorn
PyJWT
cryptography
python-dotenv
pydantic